Refactor: Add logback.xml to persist logs in directory
