+++ src/main/java/movies/movie/entity/Movie.java
@@ -1,17 +1,21 @@
-package movies.customer;
+package movies.movie.entity;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import movies.movie.type.MovieType;
 
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class Customer {
+public class Movie {
     @JsonProperty("id")
     private Long id;
 
-    @JsonProperty("name")
-    private String name;
+    @JsonProperty("title")
+    private String title;
+
+    @JsonProperty("type")
+    private MovieType type;
 }
