+++ src/main/resources/customers.json
@@ -1,14 +1,1 @@
-[
-  {
-    "id": 1,
-    "name": "Chris"
-  },
-  {
-    "id": 2,
-    "name": "John"
-  },
-  {
-    "id": 3,
-    "name": "Jane"
-  }
-]
+[{"id":1,"name":"Chris"},{"id":2,"name":"John"},{"id":3,"name":"Jane"}]
